requirements.txt
appdirs==1.4.4
CacheControl==0.13.1
certauth==1.3.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
colorama==0.4.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
digikey-api==1.0.0
distlib==0.3.7
distro==1.8.0
fake-useragent==1.2.1
filelock==3.12.2
html5lib==1.1
idna==3.4
inflection==0.5.1
inventree==0.12.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pathlib==1.0.1
pep517==0.13.0
progress==1.6
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
pytoml==0.1.21
requests==2.31.0
requests-file==1.5.1
retrying==1.3.4
schematics==2.1.1
six==1.16.0
tldextract==3.4.4
uritemplate==4.1.1
urllib3==2.0.4
webencodings==0.5.1
python-dateutil==2.8.2